Start moving to new BlockTree type
BlockTree provides references to the entrance and exit blocks, as well as
to the block where the sliced line is inserted. The intent is to make it
possible to start working from the sliced line up the tree, as well as to
make it convinient to store all the basic blocks for a function without
having to traverse the tree to find the ends. It also eases the current
refactoring process by being a different type which still holds the same
information, so it's easier to thread through things without getting
mixed up about what it represents (i.e., just passing the target basic
block wouldn't crash because of the types, but it still wouldn't work
right).