Move Conditional tags and add StartConditional tags
It's convenient to add a tag to the basic blocks on both ends of
conditional structures. This allows us to use top-down traversal for
testing, while still doing actual conversion bottom-up. Top-down is
simpler with the existing testing infrastructure, since it's already
written that way; I don't currently see any reason it won't work, so I'm
keeping it.