Finish switching to inverted basic block traversal
This handles most of the issues with ensuring we actually reach the code
we're interested in. By starting at the block we want to reach, we
ensure that we don't go down paths which we don't want to exit through.
(I.e., we don't want to take a path which returns from a conditional,
because we want to make sure to end up in the correct branch.)

Tests now pass