Add handling for generating expressions from no SSA tree
If no part of a tree is kept, the expression will just be boolean True
for now. This will of course be satisfiable. An empty tree implies any
argument value will work, so we will handle this case just as we would
any other in which not all the arguments are constrained.