Add implementations of Eq and NotEq
They're not pretty, although I believe they are correct.

Equality and inequality are awkward, because values of different types
may be compared, and should be considered not-equal. Normally, the
`else` of an `if` should be the inverse of the condition only, and the
types should be kept the same. But with equality, a valid reason for
two things to be unequal is that they have different types. So we want
to not restrict the variable to one type when dealing with an inverse
equality.

Ultimately, we can both do some basic simplification when building the
expressions and we can potentially handle comparisons with `Any`
values cleverly (i.e., require the variable to be equal to the `Any`
wrapped value; the inverse of this is exactly as desired).