Expose `TypeBuilder` `datatype` field
This allows us to customize the resulting datatype before creating it.

We also add the ability to choose a custom name for the resulting
type, and make the default name be different for each `TypeBuilder`
instance. This ensures that tests don't interfere with each other,
since we currently share a single context across all the Z3 things we
use.