Add support for `require_matching` to `MoreMagic`
This allows a function to be marked as requiring all its arguments to
be of the same sort (modulo the inheritance hierarchy). This allows
easier construction of general functions which only make sense on
same-sorted arguments but are defined for all sort types.

Also, a better registration technique for `MoreMagic`.