Correctly handle attributes in `variable_expander`
They hadn't been being handled right; we were only not expanding
attributes that were in `MAGIC_FUNCS`, not the calls on them. Now we
don't prevent expansion of the variables, but we do prevent expansion
of the call. In general, the call doesn't make sense unless the
variable is already defined somewhere, so it shouldn't be a problem
that in theory a call could be unexpanded while the variable it is
made on is expanded to an `Any` constant.