Return None from new slicer, rather than raising a RuntimeError
When the desired line cannot be found, we need to let the caller
know. A `RuntimeError` is not specific enough anyway, and returning
None lets us ignore this error easily. This is important because one
potential cause of this is an attempt to create a test case on dead
code. When we're generating tests for every line of code, we may try
to generate tests for lines which are dead, and we will want to just
ignore such lines once they are determined to be dead.