Handle store searching more carefully
We were just overwriting the old max_store without actually doing the
`max` on it. Also, if there is only the first store, it might be named
`store` with nothing more. We want to make sure to use the correct key
for the store, so we're storing the actual key for the max store.