Add support for self-referential `Pair`s
To do this, we construct and cache an empty `Pair`, then mutate it as
we determine the arguments. This lets us get a reference to the `Pair`
if we need it for one of the arguments. We have to cache and mutate,
rather than caching the lookup, because the lookup returns a Z3
`Pair`, and we need to reference the Python `Pair` that it will be
transformed into.