Add alternative names for `Pair` elements
There is a possibility that this may cause issues for code which
refers to the pair accessors as `left` and `first`, but since that
seems unlikely, I don't expect it to be a problem.

Files which actually use these names are not valid inputs for
`testbuilder`, but if rewrite rules are used to make them have
acceptable names for `testbuilder`, this allows them to still be run
in the test cases with their old names.