Add check that we do not exclude all sorts
When constructing a new `VariableTypeUnion`, it is possible we have
added incorrect type restrictions, which cause all the expressions to
be excluded, making a `TypeUnion` with no expressions. In order to
more easily catch such a bug, this assert requires that at least one
sort be present in the filtered set of sorts.