Use custom function to assign sorts to `Magic` results
Because the results may sometimes be either a `Nil`, which is an
instance of `Any`, or an actual `Any` constant, we need to have some
way to distinguish those cases. Not having any information about the
`Any` type within `Magic`, we use an external function to handle the
distinguishing for us. In the case at hand, we are using a class with
`__call__` defined, rather than a builtin function.