Simplify constraints passed into a ConstrainedExpression
This allows us to make sure that they do not have constant `True`
values. If a constraint must always be true, there is no reason it
should exist.

There is no good way to handle a constraint which can never be true,
so we will just include the `False` in the constraints and let it be
dealt with later.