Add support for floor division
Only works on integers for now. This is important, because the Python
semantics of float floor division require rounding the result toward
negative infinity, which doesn't seem to be an operation which is
available for Z3 reals. It _is_ available for Z3 FP values, but we're
not using those right now.