Add all imports at top of generated test file only
This avoids the redundancy of putting an import before each test
case. To do this, we have to add the imports to the test case code
before running it in `autogen_test`.

We had been doing this, but I didn't realize that when I added the
import lines to the individual test cases.