Do not attempt to write to `/dev/null`
Since we require that the file we're writing to not already exist,
`/dev/null` will cause a failure. But since we'd only throw away the
data written to it anyway, this throws the data away cleanly.

This is useful when checking if testbuilder is able to handle a
particular input file, without wanting the resulting tests.