Add conversion function to `testbuilder.Pair`
This allows us to ensure that the pairs we compare are both of the
same type. We convert the result `Pair` to the `Pair` from
`testbuilder` which we use as input.

This also adds support for passing in globals when creating a pair
from a `pairnet`. This allows expressions to be evaluated in the
context in which they are defined as strings, rather than in the
context of the inside of the `pairnet` function.