Add config to Makefile to force always running tests in parallel
For some reason, pytest sometimes doesn't seem to behave as it should
when all tests have passed, and the `PYTEST_FAST` run doesn't skip
everything but instead tries to run everything. This now allows us to
manually configure it to never use the `PYTEST_FAST` version if we
don't want it.